@@ -9,22 +9,22 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "illuminate/validation": "5.1.*",
-        "illuminate/events": "5.1.*",
-        "illuminate/routing": "5.1.*",
-        "illuminate/contracts": "5.1.*",
-        "illuminate/support": "5.1.*",
-        "congraph/contracts": "self.version"
+        "php": "^8.1.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/validation": "^10.0",
+        "illuminate/events": "^10.0",
+        "illuminate/routing": "^10.0",
+        "illuminate/support": "^10.0",
+        "congraph/contracts": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.5.*",
-        "orchestra/testbench": "3.1.*"
+        "phpunit/phpunit": "^10.0",
+        "orchestra/testbench": "^8.0"
     },
     "autoload": {
         "psr-4": {
             "Congraph\\Core\\": ""
         }
     },
-    "minimum-stability": "dev"
+    "minimum-stability": "stable"
 }
